backend/requirements.txt
fastapi
uvicorn[standard]
apscheduler
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
cod-api==2.0.1
httpx==0.23.0
psycopg2-binary
python-multipart
africastalking
mpesa
